Add files via upload
Linux Filesystem and Directory Permissions
Phoenix Corp runs an e-commerce website. The company’s system has a lot of data to be stored for each of its departments: Sales, HR, IT, and Customer Relationship. The company's systems are accessed by different users of other departments so the official data can be easily modified. Last week, all available space on the shared device was consumed, which affected normal operations of all departments.

To prevent this from happening in the future, the IT team has decided to partition the storage space, so that each department can be allocated dedicated storage which cannot be consumed by other departments. As a system administrator, you must implement the solution.

Requirements
Every department requires a separate disk partition mounted to a specific data directory to store their important files and information.
Departments like HR and Sales would need ext4 filesystems.
Department’s data directories should be present within the /mnt path.
The HR, Sales, IT, and Customer Relations are headed by John, Sara, Emma, and James respectively, who have complete access to their department's directory.
Employees of one department should not be able to access data from other departments.
Allocating group ownership on the directory will help ease the administrative burden in the future. Therefore, create a group for each department (for example hr_group) and provide ownership of directories to the group as well.

Tasks to complete
Create partitions, assign the filesystem, and mount them to directories as per the requirements listed in the table above. Partitions should remain mounted persistently even after reboot.
Create the four department groups as per the details given in the table.
Create the four users and add them to the specific department group as listed in the table.
Set the correct user and group owners for the data directories.
Switch to the department user's login and create department-specific files in those data directories.
Verify that only the employees of that specific department are allowed to manage its data.
Note: Partitions should remain mounted persistently even after reboot.